+++ src/main/java/com/github/cowwoc/tokenbucket/Bucket.java
@@ -4,7 +4,9 @@ import com.github.cowwoc.requirements.annotation.CheckReturnValue;
 import com.github.cowwoc.tokenbucket.Limit.ConsumptionSimulation;
 import com.github.cowwoc.tokenbucket.internal.AbstractContainer;
 import com.github.cowwoc.tokenbucket.internal.CloseableLock;
+import com.github.cowwoc.tokenbucket.internal.ContainerSecrets;
 import com.github.cowwoc.tokenbucket.internal.ReadWriteLockAsResource;
+import com.github.cowwoc.tokenbucket.internal.SharedSecrets;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -29,6 +31,7 @@ import static com.github.cowwoc.requirements.DefaultRequirements.requireThat;
  */
 public final class Bucket extends AbstractContainer
 {
+	private static final ContainerSecrets CONTAINER_SECRETS = SharedSecrets.INSTANCE.containerSecrets;
 	private final Logger log = LoggerFactory.getLogger(Bucket.class);
 	private Set<Limit> limits;
 
@@ -530,7 +533,7 @@ public final class Bucket extends AbstractContainer
 				return;
 			try (CloseableLock ignored = lock.writeLock())
 			{
-				parent.updateChild(Bucket.this, () ->
+				CONTAINER_SECRETS.updateChild(parent, Bucket.this, () ->
 				{
 					Bucket.this.limits = Set.copyOf(limits);
 					Bucket.this.userData = userData;